Stable-Diffusion-Aerial/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
safetensors>=0.4.0
certifi
